refactor(augmentation): move augmentation functions to training_utils.py
- Moved `augment_dataset`, `generate_negative_samples`, `generate_artificial_vessels`, and `gen_negative_masks` from `utils.py` to `training_utils.py`.
- Removed redundant imports in `training_utils.py` and adjusted code accordingly.
- Simplified module structure by consolidating related functions to improve maintainability and organization.
